4307 - Add inner preDiffApply functions
