+++ src/client/components/DiffDOM/types.ts
@@ -2,3 +2,11 @@ export type DiffDOMProps = {
   current: string
   prev: string
 }
+
+export type DiffInfo = {
+  diff: {
+    action: 'replaceElement' | 'addElement'
+    [key: string]: any
+  }
+  [key: string]: any
+}
